# openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/teacher/offerings:
    post:
      tags:
      - teacher-controller
      operationId: createOffering
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OfferingResponse"
  /api/teacher/offerings/{offeringId}/sessions:
    post:
      tags:
      - teacher-controller
      operationId: addSessions
      parameters:
      - name: offeringId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AddSessionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OfferingResponse"
  /api/parent/bookings:
    post:
      tags:
      - parent-controller
      operationId: bookOffering
      parameters:
      - name: timezone
        in: query
        required: false
        schema:
          type: string
          default: UTC
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookOfferingRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
  /api/teacher/{teacherId}/offerings:
    get:
      tags:
      - teacher-controller
      operationId: getTeacherOfferings
      parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: timezone
        in: query
        required: false
        schema:
          type: string
          default: UTC
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfferingResponse"
  /api/parent/{parentId}/bookings:
    get:
      tags:
      - parent-controller
      operationId: getBookings
      parameters:
      - name: parentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: timezone
        in: query
        required: false
        schema:
          type: string
          default: UTC
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
  /api/parent/offerings:
    get:
      tags:
      - parent-controller
      operationId: getAvailableOfferings
      parameters:
      - name: timezone
        in: query
        required: false
        schema:
          type: string
          default: UTC
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfferingResponse"
components:
  schemas:
    CreateOfferingRequest:
      type: object
      properties:
        courseTitle:
          type: string
          minLength: 1
        offeringName:
          type: string
          minLength: 1
        teacherId:
          type: string
          format: uuid
        teacherTimezone:
          type: string
          minLength: 1
      required:
      - courseTitle
      - offeringName
      - teacherId
      - teacherTimezone
    OfferingResponse:
      type: object
      properties:
        offeringId:
          type: string
          format: uuid
        courseTitle:
          type: string
        offeringName:
          type: string
        teacherId:
          type: string
          format: uuid
        teacherTimezone:
          type: string
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionResponse"
    SessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    AddSessionRequest:
      type: object
      properties:
        startAtLocal:
          type: string
          format: date-time
        endAtLocal:
          type: string
          format: date-time
      required:
      - endAtLocal
      - startAtLocal
    BookOfferingRequest:
      type: object
      properties:
        parentId:
          type: string
          format: uuid
        offeringId:
          type: string
          format: uuid
      required:
      - offeringId
      - parentId
    BookingResponse:
      type: object
      properties:
        bookingId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        offeringId:
          type: string
          format: uuid
        courseTitle:
          type: string
        offeringName:
          type: string
        bookedAt:
          type: string
          format: date-time
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionResponse"