# src/main/java/com/undoschool/booking/dto/AddSessionRequest.java
package com.undoschool.booking.dto;

import jakarta.validation.constraints.NotNull;

import java.time.LocalDateTime;

public record AddSessionRequest(
        @NotNull LocalDateTime startAtLocal,
        @NotNull LocalDateTime endAtLocal
) {
}